requirements.txt
# 5G-Advanced Network Optimizer - Dependencies

# Core Framework
crewai==0.86.0
crewai-tools>=0.17.0

# LLM Providers
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0

# Web Interface
streamlit>=1.40.0
plotly>=5.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.9.0

# Data handling
pandas>=2.2.0
numpy>=2.0.0
